+++ Changelog.md
@@ -19,6 +19,7 @@
 - Fix layers rendering flipped in slice preview  ([#20](https://github.com/connorslade/mslicer/issues/20))
 - Use standard unit (mm/min) for lift/retract speeds and slice height ([#22](https://github.com/connorslade/mslicer/issues/22))
 - Replace about panel with a first time popup that links to the documentation
+- Show individual model volume (before slicing)
 
 ## v0.5.0 &mdash; February 21st, 2026